ci: update config and packer
* use the new `build.sh`
* include version on packed archive

+++ .ci/pack.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
-# This script expects current working dir to be the root of repo.
-
 # This script packs files for end user, who can starts writing thesis right away.
-# To make user easy to start up, the pack should be as small as possible.
+# Only includes essential files and compiled results.
 
-${TEMP_DIR:=$(mktemp -d || echo .pack)/nuaathesis}
+# This script should run under the root of repo
+if [ ! -e .ci ]; then
+    echo "Please run under the repo root" 1>&2
+    exit 1;
+fi
+
+TEMP_ROOT=$(mktemp -d || echo .pack)
+DIR_NAME=nuaathesis-$(git rev-parse --short HEAD || echo unknown)
+TEMP_DIR="$TEMP_ROOT"/"$DIR_NAME"
 
 mkdir -p "$TEMP_DIR"
 git --git-dir=${PWD}/.git -C "$TEMP_DIR" reset --hard
@@ -14,11 +20,14 @@ git --git-dir=${PWD}/.git -C "$TEMP_DIR" reset --hard
 cp ${PWD}/nuaathesis.cls "$TEMP_DIR"
 find . -not -path '*/\.*' -type f \( -name '*.pdf' \) -exec cp --parents '{}' $TEMP_DIR \;
 find "$TEMP_DIR" -type d \( -name 'demo_*' \) -exec make -C '{}' prepare \;
-# curl -o "$TEMP_DIR"/.gitignore https://raw.githubusercontent.com/github/gitignore/master/TeX.gitignore
+
+# replace .gitignore
+rm "$TEMP_DIR"/.gitignore
+curl -o "$TEMP_DIR"/.gitignore https://raw.githubusercontent.com/github/gitignore/master/TeX.gitignore
 
 # pack
-tar czf nuaathesis.tar.gz --directory="$TEMP_DIR/.." nuaathesis
-7z a nuaathesis.zip "$TEMP_DIR"
+tar czf "$DIR_NAME".tar.gz --directory="$TEMP_ROOT" "$DIR_NAME"
+7z a "$DIR_NAME".zip "$TEMP_DIR"
 
 # clean up
-rm -rf "$TEMP_DIR"
+rm -rf "$TEMP_ROOT"
